拆解 Factory
