+++ src/factory.tsx
@@ -2,61 +2,20 @@ import React from 'react';
 import qs from 'qs';
 import {RendererStore, IRendererStore, IIRendererStore} from './store/index';
 import {getEnv, destroy} from 'mobx-state-tree';
-import {Location} from 'history';
 import {wrapFetcher} from './utils/api';
 import {normalizeLink} from './utils/normalizeLink';
-import {
-  createObject,
-  extendObject,
-  guid,
-  findIndex,
-  promisify,
-  anyChanged,
-  syncDataFromSuper,
-  isObjectShallowModified,
-  isEmpty,
-  autobind,
-  chainEvents
-} from './utils/helper';
-import {
-  Api,
-  fetcherResult,
-  Payload,
-  SchemaNode,
-  Schema,
-  Action,
-  RendererData
-} from './types';
+import {findIndex, promisify} from './utils/helper';
+import {Api, fetcherResult, Payload, SchemaNode, Schema, Action} from './types';
 import {observer} from 'mobx-react';
-import getExprProperties from './utils/filter-schema';
-import hoistNonReactStatic from 'hoist-non-react-statics';
-import omit from 'lodash/omit';
-import difference from 'lodash/difference';
-import isPlainObject from 'lodash/isPlainObject';
 import Scoped from './Scoped';
-import {
-  getTheme,
-  ThemeInstance,
-  ClassNamesFn,
-  ThemeContext,
-  ThemeProps
-} from './theme';
+import {getTheme, ThemeInstance, ThemeProps} from './theme';
 import find from 'lodash/find';
 import Alert from './components/Alert2';
 import {toast} from './components/Toast';
 import {alert, confirm, setRenderSchemaFn} from './components/Alert';
-import {LazyComponent} from './components';
-import ImageGallery from './components/ImageGallery';
-import {
-  TranslateFn,
-  getDefaultLocale,
-  makeTranslator,
-  LocaleContext,
-  LocaleProps
-} from './locale';
-import {SchemaCollection, SchemaObject, SchemaTpl} from './Schema';
-import {result} from 'lodash';
-import {envOverwrite} from './envOverwrite';
+import {getDefaultLocale, makeTranslator, LocaleProps} from './locale';
+import ScopedRootRenderer, {RootRenderProps} from './Root';
+import {HocStoreFactory} from './WithStore';
 
 export interface TestFunc {
   (
@@ -148,10 +107,6 @@ export interface RendererProps extends ThemeProps, LocaleProps {
   [propName: string]: any;
 }
 
-export interface renderChildProps extends Partial<RendererProps> {
-  env: RendererEnv;
-}
-
 export type RendererComponent = React.ComponentType<RendererProps> & {
   propsList?: Array<any>;
 };
@@ -163,12 +118,6 @@ export interface RendererConfig extends RendererBasicConfig {
 
 export interface RenderSchemaFilter {
   (schema: Schema, renderer: RendererConfig, props?: any): Schema;
-}
-
-export interface RootRenderProps {
-  location?: Location;
-  theme?: string;
-  [propName: string]: any;
 }
 
 export interface RenderOptions {
@@ -203,8 +152,6 @@ export interface fetcherConfig {
   data?: any;
   config?: any;
 }
-
-export type ReactElement = React.ReactNode[] | JSX.Element | null | false;
 
 const renderers: Array<RendererConfig> = [];
 const rendererNames: Array<string> = [];
@@ -291,699 +238,13 @@ export function unRegisterRenderer(config: RendererConfig | string) {
   cache = {};
 }
 
-export function renderChildren(
-  prefix: string,
-  node: SchemaNode,
-  props: renderChildProps
-): ReactElement {
-  if (Array.isArray(node)) {
-    return node.map((node, index) =>
-      renderChild(`${prefix}/${index}`, node, {
-        ...props,
-        key: `${props.key ? `${props.key}-` : ''}${index}`
-      })
-    );
-  }
-
-  return renderChild(prefix, node, props);
-}
-
-export function renderChild(
-  prefix: string,
-  node: SchemaNode,
-  props: renderChildProps
-): ReactElement {
-  if (Array.isArray(node)) {
-    return renderChildren(prefix, node, props);
-  }
-
-  const typeofnode = typeof node;
-  let schema: Schema =
-    typeofnode === 'string' || typeofnode === 'number'
-      ? {type: 'tpl', tpl: String(node)}
-      : (node as Schema);
-  const detectData =
-    schema &&
-    (schema.detectField === '&' ? props : props[schema.detectField || 'data']);
-  const exprProps = detectData
-    ? getExprProperties(schema, detectData, undefined, props)
-    : null;
-
-  if (
-    exprProps &&
-    (exprProps.hidden ||
-      exprProps.visible === false ||
-      schema.hidden ||
-      schema.visible === false ||
-      props.hidden ||
-      props.visible === false)
-  ) {
-    return null;
-  }
-
-  const transform = props.propsTransform;
-
-  if (transform) {
-    // @ts-ignore
-    delete props.propsTransform;
-    props = transform(props);
-  }
-
-  return (
-    <SchemaRenderer
-      {...props}
-      {...exprProps}
-      schema={schema}
-      $path={`${prefix ? `${prefix}/` : ''}${(schema && schema.type) || ''}`}
-    />
-  );
-}
-
-export interface RootRendererProps {
-  schema: SchemaNode;
-  rootStore: IRendererStore;
-  env: RendererEnv;
-  theme: string;
-  pathPrefix?: string;
-  locale?: string;
-  translate?: TranslateFn;
-  [propName: string]: any;
-}
-
-export const RootStoreContext = React.createContext<IRendererStore>(
-  undefined as any
-);
-
-export class RootRenderer extends React.Component<RootRendererProps> {
-  state = {
-    error: null,
-    errorInfo: null
-  };
-
-  componentDidCatch(error: any, errorInfo: any) {
-    console.error(error);
-    this.setState({
-      error: error,
-      errorInfo: errorInfo
-    });
-  }
-
-  @autobind
-  resolveDefinitions(name: string) {
-    const definitions = (this.props.schema as Schema).definitions;
-    if (!name || isEmpty(definitions)) {
-      return {};
-    }
-    return definitions && definitions[name];
-  }
-
-  render() {
-    const {error, errorInfo} = this.state;
-    if (errorInfo) {
-      return errorRenderer(error, errorInfo);
-    }
-    const {
-      schema,
-      rootStore,
-      env,
-      pathPrefix,
-      location,
-      data,
-      locale,
-      translate,
-      ...rest
-    } = this.props;
-
-    const theme = env.theme;
-    const query =
-      (location && location.query) ||
-      (location && location.search && qs.parse(location.search.substring(1))) ||
-      (window.location.search && qs.parse(window.location.search.substring(1)));
-
-    const finalData = query
-      ? createObject(
-          {
-            ...(data && data.__super ? data.__super : null),
-            ...query,
-            __query: query
-          },
-          data
-        )
-      : data;
-
-    // 根据环境覆盖 schema，这个要在最前面做，不然就无法覆盖 validations
-    envOverwrite(schema, locale);
-
-    return (
-      <RootStoreContext.Provider value={rootStore}>
-        <ThemeContext.Provider value={this.props.theme || 'default'}>
-          <LocaleContext.Provider value={this.props.locale!}>
-            <ImageGallery modalContainer={env.getModalContainer}>
-              {
-                renderChild(
-                  pathPrefix || '',
-                  isPlainObject(schema)
-                    ? {
-                        type: 'page',
-                        ...(schema as any)
-                      }
-                    : schema,
-                  {
-                    ...rest,
-                    resolveDefinitions: this.resolveDefinitions,
-                    location: location,
-                    data: finalData,
-                    env,
-                    classnames: theme.classnames,
-                    classPrefix: theme.classPrefix,
-                    locale,
-                    translate
-                  }
-                ) as JSX.Element
-              }
-            </ImageGallery>
-          </LocaleContext.Provider>
-        </ThemeContext.Provider>
-      </RootStoreContext.Provider>
-    );
-  }
-}
-
-export const ScopedRootRenderer = Scoped(RootRenderer);
-
-interface SchemaRendererProps extends Partial<RendererProps> {
-  schema: Schema;
-  $path: string;
-  env: RendererEnv;
-}
-
-const defaultOmitList = [
-  'type',
-  'name',
-  '$ref',
-  'className',
-  'data',
-  'children',
-  'ref',
-  'visible',
-  'visibleOn',
-  'hidden',
-  'hiddenOn',
-  'disabled',
-  'disabledOn',
-  'component',
-  'detectField',
-  'required',
-  'requiredOn',
-  'syncSuperStore'
-];
-
-class SchemaRenderer extends React.Component<SchemaRendererProps, any> {
-  static displayName: string = 'Renderer';
-
-  renderer: RendererConfig | null;
-  ref: any;
-  schema: any;
-  path: string;
-
-  constructor(props: SchemaRendererProps) {
-    super(props);
-    this.refFn = this.refFn.bind(this);
-    this.renderChild = this.renderChild.bind(this);
-    this.reRender = this.reRender.bind(this);
-  }
-
-  componentWillMount() {
-    this.resolveRenderer(this.props);
-  }
-
-  componentWillReceiveProps(nextProps: SchemaRendererProps) {
-    const props = this.props;
-
-    if (
-      props.schema &&
-      nextProps.schema &&
-      (props.schema.type !== nextProps.schema.type ||
-        props.schema.$$id !== nextProps.schema.$$id)
-    ) {
-      this.resolveRenderer(nextProps);
-    }
-  }
-
-  // 限制：只有 schema 除外的 props 变化，或者 schema 里面的某个成员值发生变化才更新。
-  shouldComponentUpdate(nextProps: SchemaRendererProps) {
-    const props = this.props;
-    const list: Array<string> = difference(Object.keys(nextProps), [
-      'schema',
-      'scope'
-    ]);
-
-    if (
-      difference(Object.keys(props), ['schema', 'scope']).length !==
-        list.length ||
-      anyChanged(list, this.props, nextProps)
-    ) {
-      return true;
-    } else {
-      const list: Array<string> = Object.keys(nextProps.schema);
-
-      if (
-        Object.keys(props.schema).length !== list.length ||
-        anyChanged(list, props.schema, nextProps.schema)
-      ) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  resolveRenderer(props: SchemaRendererProps, skipResolve = false): any {
-    let schema = props.schema;
-    let path = props.$path;
-
-    if (schema && schema.$ref) {
-      schema = {
-        ...props.resolveDefinitions(schema.$ref),
-        ...schema
-      };
-
-      path = path.replace(/(?!.*\/).*/, schema.type);
-    }
-
-    if (!skipResolve) {
-      const rendererResolver = props.env.rendererResolver || resolveRenderer;
-      this.renderer = rendererResolver(path, schema, props);
-    }
-
-    return {path, schema};
-  }
-
-  getWrappedInstance() {
-    return this.ref;
-  }
-
-  refFn(ref: any) {
-    this.ref = ref;
-  }
-
-  renderChild(
-    region: string,
-    node?: SchemaNode,
-    subProps: {
-      data?: object;
-      [propName: string]: any;
-    } = {}
-  ) {
-    let {schema, $path, env, ...rest} = this.props;
-
-    if (schema && schema.$ref) {
-      const result = this.resolveRenderer(this.props, true);
-      schema = result.schema;
-      $path = result.path;
-    }
-
-    const omitList = defaultOmitList.concat();
-    if (this.renderer) {
-      const Component = this.renderer.component;
-      Component.propsList &&
-        omitList.push.apply(omitList, Component.propsList as Array<string>);
-    }
-
-    return renderChild(`${$path}${region ? `/${region}` : ''}`, node || '', {
-      ...omit(rest, omitList),
-      ...subProps,
-      data: subProps.data || rest.data,
-      env: env
-    });
-  }
-
-  reRender() {
-    this.resolveRenderer(this.props);
-    this.forceUpdate();
-  }
-
-  render(): JSX.Element | null {
-    let {$path, schema, ...rest} = this.props;
-
-    if (schema === null) {
-      return null;
-    }
-
-    if (schema.$ref) {
-      const result = this.resolveRenderer(this.props, true);
-      schema = result.schema;
-      $path = result.path;
-    }
-
-    const theme = this.props.env.theme;
-
-    if (Array.isArray(schema)) {
-      return renderChildren($path, schema as any, rest) as JSX.Element;
-    } else if (schema.children) {
-      return React.isValidElement(schema.children)
-        ? schema.children
-        : (schema.children as Function)({
-            ...rest,
-            $path: $path,
-            render: this.renderChild,
-            forwardedRef: this.refFn
-          });
-    } else if (typeof schema.component === 'function') {
-      const isSFC = !(schema.component.prototype instanceof React.Component);
-      return React.createElement(schema.component as any, {
-        ...rest,
-        ...schema,
-        $path: $path,
-        ref: isSFC ? undefined : this.refFn,
-        forwardedRef: isSFC ? this.refFn : undefined,
-        render: this.renderChild
-      });
-    } else if (Object.keys(schema).length === 0) {
-      return null;
-    } else if (!this.renderer) {
-      return (
-        <LazyComponent
-          {...rest}
-          getComponent={async () => {
-            const result = await rest.env.loadRenderer(
-              schema,
-              $path,
-              this.reRender
-            );
-            if (result && typeof result === 'function') {
-              return result;
-            } else if (result && React.isValidElement(result)) {
-              return () => result;
-            }
-
-            this.reRender();
-            return () => loadRenderer(schema, $path);
-          }}
-          $path={$path}
-          retry={this.reRender}
-        />
-      );
-    }
-
-    const renderer = this.renderer as RendererConfig;
-    schema = filterSchema(schema, renderer, rest);
-    const {data: defaultData, ...restSchema} = schema;
-    const Component = renderer.component;
-
-    return (
-      <Component
-        {...theme.getRendererConfig(renderer.name)}
-        {...restSchema}
-        {...chainEvents(rest, restSchema)}
-        defaultData={defaultData}
-        $path={$path}
-        ref={this.refFn}
-        render={this.renderChild}
-      />
-    );
-  }
-}
-
-export function HocStoreFactory(renderer: {
-  storeType: string;
-  extendsData?: boolean;
-  shouldSyncSuperStore?: (
-    store: any,
-    props: any,
-    prevProps: any
-  ) => boolean | undefined;
-}): any {
-  return function <T extends React.ComponentType<RendererProps>>(Component: T) {
-    type Props = Omit<
-      RendererProps,
-      'store' | 'data' | 'dataUpdatedAt' | 'scope'
-    > & {
-      store?: IIRendererStore;
-      data?: RendererData;
-      scope?: RendererData;
-    };
-
-    @observer
-    class StoreFactory extends React.Component<Props> {
-      static displayName = `WithStore(${
-        Component.displayName || Component.name
-      })`;
-      static ComposedComponent = Component;
-      static contextType = RootStoreContext;
-      store: IIRendererStore;
-      context!: React.ContextType<typeof RootStoreContext>;
-      ref: any;
-
-      getWrappedInstance() {
-        return this.ref;
-      }
-
-      refFn(ref: any) {
-        this.ref = ref;
-      }
-
-      formatData(data: any): object {
-        if (Array.isArray(data)) {
-          return {
-            items: data
-          };
-        }
-
-        return data as object;
-      }
-
-      componentWillMount() {
-        const rootStore = this.context;
-        this.renderChild = this.renderChild.bind(this);
-        this.refFn = this.refFn.bind(this);
-
-        const store = rootStore.addStore({
-          id: guid(),
-          path: this.props.$path,
-          storeType: renderer.storeType,
-          parentId: this.props.store ? this.props.store.id : ''
-        }) as IIRendererStore;
-        this.store = store;
-
-        if (renderer.extendsData === false) {
-          store.initData(
-            createObject(
-              (this.props.data as any)
-                ? (this.props.data as any).__super
-                : null,
-              {
-                ...this.formatData(this.props.defaultData),
-                ...this.formatData(this.props.data)
-              }
-            )
-          );
-        } else if (
-          this.props.scope ||
-          (this.props.data && (this.props.data as any).__super)
-        ) {
-          if (this.props.store && this.props.data === this.props.store.data) {
-            store.initData(
-              createObject(this.props.store.data, {
-                ...this.formatData(this.props.defaultData)
-              })
-            );
-          } else {
-            store.initData(
-              createObject(
-                (this.props.data as any).__super || this.props.scope,
-                {
-                  ...this.formatData(this.props.defaultData),
-                  ...this.formatData(this.props.data)
-                }
-              )
-            );
-          }
-        } else {
-          store.initData({
-            ...this.formatData(this.props.defaultData),
-            ...this.formatData(this.props.data)
-          });
-        }
-      }
-
-      componentWillReceiveProps(nextProps: RendererProps) {
-        const props = this.props;
-        const store = this.store;
-        const shouldSync = renderer.shouldSyncSuperStore?.(
-          store,
-          nextProps,
-          props
-        );
-
-        if (shouldSync === false) {
-          return;
-        }
-
-        if (renderer.extendsData === false) {
-          if (
-            shouldSync === true ||
-            props.defaultData !== nextProps.defaultData ||
-            isObjectShallowModified(props.data, nextProps.data) ||
-            //
-            // 特殊处理 CRUD。
-            // CRUD 中 toolbar 里面的 data 是空对象，但是 __super 会不一样
-            (nextProps.data &&
-              props.data &&
-              nextProps.data.__super !== props.data.__super)
-          ) {
-            store.initData(
-              extendObject(nextProps.data, {
-                ...(store.hasRemoteData ? store.data : null), // todo 只保留 remote 数据
-                ...this.formatData(nextProps.defaultData),
-                ...this.formatData(nextProps.data)
-              })
-            );
-          }
-        } else if (
-          shouldSync === true ||
-          isObjectShallowModified(props.data, nextProps.data)
-        ) {
-          if (nextProps.store && nextProps.store.data === nextProps.data) {
-            store.initData(
-              createObject(
-                nextProps.store.data,
-                nextProps.syncSuperStore === false
-                  ? {
-                      ...store.data
-                    }
-                  : syncDataFromSuper(
-                      store.data,
-                      nextProps.store.data,
-                      props.scope,
-                      store,
-                      nextProps.syncSuperStore === true
-                    )
-              )
-            );
-          } else if (nextProps.data && (nextProps.data as any).__super) {
-            store.initData(extendObject(nextProps.data));
-          } else {
-            store.initData(createObject(nextProps.scope, nextProps.data));
-          }
-        } else if (
-          (shouldSync === true ||
-            !nextProps.store ||
-            nextProps.data !== nextProps.store.data) &&
-          nextProps.data &&
-          nextProps.data.__super
-        ) {
-          // 这个用法很少，当 data.__super 值发生变化时，更新 store.data
-          (!props.data ||
-            isObjectShallowModified(
-              nextProps.data.__super,
-              props.data.__super,
-              false
-            )) &&
-            // nextProps.data.__super !== props.data.__super) &&
-            store.initData(
-              createObject(nextProps.data.__super, {
-                ...nextProps.data,
-                ...store.data
-              }),
-
-              store.storeType === 'FormStore' &&
-                props.store?.storeType === 'CRUDStore'
-            );
-        } else if (
-          nextProps.scope &&
-          nextProps.data === nextProps.store!.data &&
-          (shouldSync === true || props.data !== nextProps.data)
-        ) {
-          store.initData(
-            createObject(nextProps.scope, {
-              // ...nextProps.data,
-              ...store.data
-            })
-          );
-        }
-      }
-
-      componentWillUnmount() {
-        const rootStore = this.context as IRendererStore;
-        const store = this.store;
-        rootStore.removeStore(store);
-
-        // @ts-ignore
-        delete this.store;
-      }
-
-      renderChild(
-        region: string,
-        node: SchemaNode,
-        subProps: {
-          data?: object;
-          [propName: string]: any;
-        } = {}
-      ) {
-        let {render} = this.props;
-
-        return render(region, node, {
-          data: this.store.data,
-          dataUpdatedAt: this.store.updatedAt,
-          ...subProps,
-          scope: this.store.data,
-          store: this.store
-        });
-      }
-
-      render() {
-        const {detectField, ...rest} = this.props;
-
-        let exprProps: any = {};
-        if (!detectField || detectField === 'data') {
-          exprProps = getExprProperties(rest, this.store.data, undefined, rest);
-
-          if (exprProps.hidden || exprProps.visible === false) {
-            return null;
-          }
-        }
-
-        return (
-          <Component
-            {
-              ...(rest as any) /* todo */
-            }
-            {...exprProps}
-            ref={this.refFn}
-            data={this.store.data}
-            dataUpdatedAt={this.store.updatedAt}
-            store={this.store}
-            scope={this.store.data}
-            render={this.renderChild}
-          />
-        );
-      }
-    }
-    hoistNonReactStatic(StoreFactory, Component);
-
-    return StoreFactory;
-  };
-}
-
-function loadRenderer(schema: Schema, path: string) {
+export function loadRenderer(schema: Schema, path: string) {
   return (
     <Alert level="danger">
       <p>Error: 找不到对应的渲染器</p>
       <p>Path: {path}</p>
       <pre>
         <code>{JSON.stringify(schema, null, 2)}</code>
-      </pre>
-    </Alert>
-  );
-}
-
-function errorRenderer(error: any, errorInfo: any) {
-  return (
-    <Alert level="danger">
-      <p>{error && error.toString()}</p>
-      <pre>
-        <code>{errorInfo.componentStack}</code>
       </pre>
     </Alert>
   );
@@ -1143,7 +404,7 @@ export function clearStoresCache(
 }
 
 // 当然也可以直接这样更新。
-// 主要是有时候第一创建的时候并没有准备多少接口，
+// 主要是有时候第一次创建的时候并没有准备多少接口，
 // 可以后续补充点，比如 amis 自己实现的，prompt 里面的表单。
 export function updateEnv(options: Partial<RenderOptions>, session = 'global') {
   options = {
@@ -1171,8 +432,7 @@ export function updateEnv(options: Partial<RenderOptions>, session = 'global') {
 let cache: {[propName: string]: RendererConfig} = {};
 export function resolveRenderer(
   path: string,
-  schema?: Schema,
-  props?: any
+  schema?: Schema
 ): null | RendererConfig {
   if (cache[path]) {
     return cache[path];
@@ -1185,6 +445,7 @@ export function resolveRenderer(
   renderers.some(item => {
     let matched = false;
 
+    // 不应该搞得这么复杂的，让每个渲染器唯一 id，自己不晕别人用起来也不晕。
     if (typeof item.test === 'function') {
       matched = item.test(path, schema, resolveRenderer);
     } else if (item.test instanceof RegExp) {
@@ -1198,7 +459,8 @@ export function resolveRenderer(
     return matched;
   });
 
-  // 只能缓存纯正则表达式的后者方法中没有用到第二个参数的，因为自定义 test 函数的有可能依赖 schema 的结果
+  // 只能缓存纯正则表达式的后者方法中没有用到第二个参数的，
+  // 因为自定义 test 函数的有可能依赖 schema 的结果
   if (
     renderer !== null &&
     ((renderer as RendererConfig).test instanceof RegExp ||
@@ -1217,53 +479,6 @@ export function getRenderers() {
 
 export function getRendererByName(name: string) {
   return find(renderers, item => item.name === name);
-}
-
-export function withRootStore<
-  T extends React.ComponentType<
-    React.ComponentProps<T> & {
-      rootStore: IRendererStore;
-    }
-  >
->(ComposedComponent: T) {
-  type OuterProps = JSX.LibraryManagedAttributes<
-    T,
-    Omit<React.ComponentProps<T>, 'rootStore'>
-  >;
-
-  const result = hoistNonReactStatic(
-    class extends React.Component<OuterProps> {
-      static displayName = `WithRootStore(${
-        ComposedComponent.displayName || ComposedComponent.name
-      })`;
-      static contextType = RootStoreContext;
-      static ComposedComponent = ComposedComponent;
-
-      render() {
-        const rootStore = this.context;
-        const injectedProps: {
-          rootStore: IRendererStore;
-        } = {
-          rootStore
-        };
-
-        return (
-          <ComposedComponent
-            {...(this.props as JSX.LibraryManagedAttributes<
-              T,
-              React.ComponentProps<T>
-            >)}
-            {...injectedProps}
-          />
-        );
-      }
-    },
-    ComposedComponent
-  );
-
-  return result as typeof result & {
-    ComposedComponent: T;
-  };
 }
 
 setRenderSchemaFn((controls, value, callback, scopeRef, theme) => {
